[RND-2661] wifi: brcmfmac: Fix suspend for boards that needs power preserved
There are some boards that show issues with wifi adapters when the SDIO
bus is unpowered. On these boards, the wifi adapter doesn't come up
anymore.

This is a regression caused by commit 92cadedd9d5f ("brcmfmac: Avoid
keeping power to SDIO card unless WOWL is used"), the wifi adapter by
default is turned off on suspend and then re-probed on resume.

This introduces a new flag that can be used by boards that require the
old default behaviour in which the SDIO remains powered.

Signed-off-by: Matthias Proske <matthias.p@variscite.com>